Biggest commit of my life. Never do this.
Got the login spec working to where the login page doesn't have any console
errors. The dashboard page is throwing errors, but that's doable.

I reorganized all the emerald stuff that was scattered all over the place.

Created a default rake task that runs the cucumber tasks marked as @wip (Work in
Progress).