The "user logs in" user story is sufficiently tested.
The cucumber tests are passing. The associated controller has been tested. The
only way we could test it better is to prove through unit tests that the
currentUser service works as expected, but that's so tightly coupled to firebase
that after mocking the firebase stuff, we're not actually unit testing anything.

As such, I'm making the call to not do any unit testing on it. Testing for the
sake of testing is a worthless endeavor.